Backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib[bcrypt]
pydantic
python-multipart
openpyxl
qrcode